workout-tracker/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.3.0
numpy==1.25.2
opencv-python==4.8.0.76
mediapipe==0.10.3
google-generativeai==0.1.0
python-dotenv==1.0.0
websockets==11.0.3